clippy fixes

+++ transform-storage/src/conversions/quant/float_quant.rs
@@ -23,7 +23,7 @@ pub fn quantize_float_bits(value: f32, dtype: QuantDType) -> f32 {
 
     let log2_val = abs_val.log2();
     let exp_unbiased = log2_val.floor() as i32;
-    let exp_clamped = exp_unbiased.clamp(-(exp_bias as i32), exp_max as i32 - exp_bias as i32 - 1);
+    let exp_clamped = exp_unbiased.clamp(-exp_bias, exp_max - exp_bias - 1);
 
     let mantissa_full = abs_val / f32::exp2(exp_clamped as f32) - 1.0;
     let mantissa_steps = (1 << mantissa_bits) as f32;
